+++ hutool-core/src/main/java/cn/hutool/core/text/csv/CsvRow.java
@@ -177,7 +177,7 @@ public final class CsvRow implements List<String> {
 
 	@Override
 	public String get(int index) {
-		return index > fields.size() ? null : fields.get(index);
+		return index >= fields.size() ? null : fields.get(index);
 	}
 
 	@Override