+++ hutool-poi/src/main/java/cn/hutool/poi/excel/WorkbookUtil.java
@@ -12,7 +12,6 @@ import org.apache.poi.ss.usermodel.WorkbookFactory;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IORuntimeException;
 import cn.hutool.core.io.IoUtil;
-import cn.hutool.core.util.StrUtil;
 import cn.hutool.poi.exceptions.POIException;
 
 /**
@@ -25,50 +24,54 @@ import cn.hutool.poi.exceptions.POIException;
 public class WorkbookUtil {
 	
 	/**
-	 * 加载工作簿
+	 * 创建或加载工作簿
 	 * 
 	 * @param excelFilePath Excel文件路径，绝对路径或相对于ClassPath路径
 	 * @return {@link Workbook}
 	 * @since 3.1.1
 	 */
-	public static Workbook loadBook(String excelFilePath) {
-		return loadBook(FileUtil.file(excelFilePath), null);
+	public static Workbook createBook(String excelFilePath) {
+		return createBook(FileUtil.file(excelFilePath), null);
 	}
 
 	/**
-	 * 加载工作簿
+	 * 创建或加载工作簿
 	 * 
 	 * @param excelFile Excel文件
 	 * @return {@link Workbook}
 	 */
-	public static Workbook loadBook(File excelFile) {
-		return loadBook(excelFile, null);
+	public static Workbook createBook(File excelFile) {
+		return createBook(excelFile, null);
 	}
 
 	/**
-	 * 加载工作簿，只读模式
+	 * 创建或加载工作簿，只读模式
 	 * 
 	 * @param excelFile Excel文件
 	 * @param password Excel工作簿密码，如果无密码传{@code null}
 	 * @return {@link Workbook}
 	 */
-	public static Workbook loadBook(File excelFile, String password) {
-		return loadBook(FileUtil.getInputStream(excelFile), password, true);
+	public static Workbook createBook(File excelFile, String password) {
+		try {
+			return WorkbookFactory.create(excelFile, password);
+		} catch (Exception e) {
+			throw new POIException(e);
+		}
 	}
 
 	/**
-	 * 加载工作簿
+	 * 创建或加载工作簿
 	 * 
 	 * @param in Excel输入流
 	 * @param closeAfterRead 读取结束是否关闭流
 	 * @return {@link Workbook}
 	 */
-	public static Workbook loadBook(InputStream in, boolean closeAfterRead) {
-		return loadBook(in, null, closeAfterRead);
+	public static Workbook createBook(InputStream in, boolean closeAfterRead) {
+		return createBook(in, null, closeAfterRead);
 	}
 
 	/**
-	 * 加载工作簿
+	 * 创建或加载工作簿
 	 * 
 	 * @param in Excel输入流
 	 * @param password 密码
@@ -76,7 +79,7 @@ public class WorkbookUtil {
 	 * @return {@link Workbook}
 	 * @since 4.0.3
 	 */
-	public static Workbook loadBook(InputStream in, String password, boolean closeAfterRead) {
+	public static Workbook createBook(InputStream in, String password, boolean closeAfterRead) {
 		try {
 			return WorkbookFactory.create(IoUtil.toMarkSupportStream(in), password);
 		} catch (Exception e) {
@@ -87,33 +90,22 @@ public class WorkbookUtil {
 			}
 		}
 	}
-
+	
 	/**
 	 * 根据文件路径创建新的工作簿，文件路径
 	 * 
 	 * @param destFilePath 目标文件路径，文件可以不存在，通过扩展名判断Workbook类型
 	 * @return {@link Workbook}
-	 * @since 3.2.0
+	 * @since 4.1.0
 	 */
-	public static Workbook createBook(String destFilePath) {
+	public static Workbook createBook(boolean isXlsx) {
 		Workbook workbook;
-		if (StrUtil.endWithIgnoreCase(destFilePath, "xlsx")) {
+		if (isXlsx) {
 			workbook = new org.apache.poi.xssf.usermodel.XSSFWorkbook();
 		} else {
 			workbook = new org.apache.poi.hssf.usermodel.HSSFWorkbook();
 		}
 		return workbook;
-	}
-
-	/**
-	 * 根据文件路径创建新的工作簿
-	 * 
-	 * @param destFile 目标文件，文件可以不存在
-	 * @return {@link Workbook}
-	 * @since 3.2.0
-	 */
-	public static Workbook createBook(File destFile) {
-		return createBook(destFile.getName());
 	}
 
 	/**