+++ hutool-poi/src/main/java/cn/hutool/poi/excel/style/StyleUtil.java
@@ -1,4 +1,4 @@
-package cn.hutool.poi.excel;
+package cn.hutool.poi.excel.style;
 
 import org.apache.poi.ss.usermodel.BorderStyle;
 import org.apache.poi.ss.usermodel.Cell;